re-enabled basic colorplotting. deleted old templates.
